refactor(catalog-cli): carve backfill-owner-id into catalog_cmds (nexus-kgyoz)
Second deferred kgyoz CLI carve, mirroring the seam-3 owners pattern. Lift
the one-time RDR-137 P1.5a migration command backfill-owner-id out of
commands/catalog.py into catalog_cmds/backfill.py as a plain click command
plus a register(group) hook; commands.catalog calls _backfill_cmds.register(
catalog) alongside the owners registration. 'nx catalog backfill-owner-id'
resolves identically — name, both options, help, output, and the
service-mode guard unchanged.

Behaviour-preserving: body moved verbatim (incl. the epsilon-allow SQLite
markers); _get_catalog reached via the nexus.commands.catalog module object
so the import stays acyclic and the _get_catalog patch seam holds. No test
imports the command by symbol.

Three contract pins in test_catalog_cli.py::TestKgyozBackfillCarve:
group-registration, callback-identity-in-carved-module, and a behavioural
service-mode-guard pin that runs the command end-to-end through the group.
Guard suites green (99 passed).
