+++ src/nexus/commands/catalog.py
@@ -232,9 +232,11 @@ def catalog() -> None:
 # here (after the group exists) so ``nx catalog owners`` / ``dedupe-owners``
 # resolve identically. catalog_cmds submodules reference this module's helpers
 # lazily, so this import stays acyclic.
+from nexus.commands.catalog_cmds import backfill as _backfill_cmds  # noqa: E402 — must follow the `catalog` group definition above
 from nexus.commands.catalog_cmds import owners as _owners_cmds  # noqa: E402 — must follow the `catalog` group definition above
 
 _owners_cmds.register(catalog)
+_backfill_cmds.register(catalog)
 
 
 @catalog.command("init")
@@ -466,84 +468,6 @@ def backfill_collections_cmd(dry_run: bool) -> None:
         f"\nDone: {len(to_register)} new, "
         f"{len(already)} already registered."
     )
-
-
-@catalog.command("backfill-owner-id")
-@click.option(
-    "--dry-run/--no-dry-run",
-    default=True,
-    help="Report-only (default). Use --no-dry-run to actually update. "
-    "Matches the safe-default convention of the other Phase 6 verbs.",
-)
-@click.option(
-    "--from-documents/--no-from-documents",
-    default=True,
-    help="Use the documents-table fallback to recover owner_id for legacy "
-    "2-segment names (default). Disable to restrict the backfill to "
-    "the auto-migration's conformant-name path only.",
-)
-def backfill_owner_id_cmd(dry_run: bool, from_documents: bool) -> None:
-    """Populate ``collections.owner_id`` for empty rows (RDR-137 P1.5a).
-
-    \b
-    The CatalogStore's auto-migration handles conformant RDR-103
-    four-segment names on every DB open. This verb adds the documents-
-    table fallback that recovers owner_id for legacy 2-segment names
-    (e.g. ``knowledge__delos``) by inferring owner from documents that
-    are physically registered against the collection.
-
-    \b
-    Ambiguous rows (documents from multiple distinct owners) are
-    skipped with a warning. The auto-migration is idempotent — running
-    this with ``--from-documents=false`` is equivalent to opening any
-    CatalogStore connection.
-
-    \b
-    Examples:
-      nx catalog backfill-owner-id --dry-run
-      nx catalog backfill-owner-id --no-dry-run
-      nx catalog backfill-owner-id --no-dry-run --no-from-documents
-    """
-    from nexus.catalog.collections_owner_backfill import (  # noqa: PLC0415  — command-local import (nexus.catalog.collections_owner_backfill)
-        backfill_owner_id,
-    )
-    from nexus.catalog.factory import _is_catalog_service_mode  # noqa: PLC0415  — command-local import (nexus.catalog.factory)
-
-    if _is_catalog_service_mode():
-        raise click.ClickException(
-            "nx catalog backfill-owner-id is not supported in service mode — "
-            "this one-time migration runs against the SQLite database directly. "
-            "Run it locally before switching to service mode. "
-            "Tracked in nexus follow-on bead (gated on nexus-gmiaf.24)."
-        )
-
-    cat = _get_catalog()
-    # epsilon-allow: backfill_owner_id() requires a raw SQLite connection by
-    # contract; service mode is guarded above. This pair will remain until
-    # the backfill function is refactored to use the catalog API.
-    with cat._db:  # epsilon-allow: SQLite-only write, service mode guarded above
-        result = backfill_owner_id(
-            cat._db,  # epsilon-allow: SQLite-only write, service mode guarded above
-            include_documents_fallback=from_documents,
-            dry_run=dry_run,
-        )
-
-    verb = "would update" if dry_run else "updated"
-    click.echo(
-        f"{verb} {result.updated_from_name} via name + "
-        f"{result.updated_from_documents} via documents fallback "
-        f"(total empty: {result.total_empty})"
-    )
-    if result.skipped_ambiguous:
-        click.echo(
-            f"  skipped {result.skipped_ambiguous} ambiguous "
-            f"(multi-owner) collection(s)"
-        )
-    if result.skipped_unresolvable:
-        click.echo(
-            f"  skipped {result.skipped_unresolvable} unresolvable "
-            f"collection(s) — manual review required"
-        )
 
 
 @catalog.command("migrate-fallback")
